# api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/event-types": {
      "post": {
        "operationId": "EventTypesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventTypeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "event-types"
        ]
      },
      "get": {
        "operationId": "EventTypesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "event-types"
        ]
      }
    },
    "/event-types/{id}": {
      "get": {
        "operationId": "EventTypesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "event-types"
        ]
      },
      "patch": {
        "operationId": "EventTypesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventTypeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "event-types"
        ]
      },
      "delete": {
        "operationId": "EventTypesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "event-types"
        ]
      }
    },
...
    "/eligibility": {
      "post": {
        "operationId": "EligibilityController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEligibilityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "eligibility"
        ]
      },
      "get": {
        "operationId": "EligibilityController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "eligibility"
        ]
      }
    },
    "/eligibility/{id}": {
      "get": {
        "operationId": "EligibilityController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "eligibility"
        ]
      },
      "patch": {
        "operationId": "EligibilityController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEligibilityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "eligibility"
        ]
      },
      "delete": {
        "operationId": "EligibilityController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "eligibility"
        ]
      }
    },
...
    "/events/{id}": {
      "get": {
        "operationId": "EventsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "events"
        ]
      },
      "patch": {
        "operationId": "EventsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "events"
        ]
      },
      "delete": {
        "operationId": "EventsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "events"
        ]
      }
    },
...
    "/categories": {
      "post": {
        "operationId": "CategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "categories"
        ]
      },
      "get": {
        "operationId": "CategoriesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "categories"
        ]
      }
    },
...
    "/categories/{id}": {
      "get": {
        "operationId": "CategoriesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "categories"
        ]
      },
      "patch": {
        "operationId": "CategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "categories"
        ]
      },
      "delete": {
        "operationId": "CategoriesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "categories"
        ]
      }
    },
...
    "/card-generation-type": {
      "post": {
        "operationId": "CardGenerationController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCardGenerationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "card-generation-type"
        ]
      },
      "get": {
        "operationId": "CardGenerationController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "card-generation-type"
        ]
      }
    },
    "/card-generation-type/{id}": {
      "get": {
        "operationId": "CardGenerationController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "card-generation-type"
        ]
      },
      "patch": {
        "operationId": "CardGenerationController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCardGenerationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "card-generation-type"
        ]
      },
      "delete": {
        "operationId": "CardGenerationController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "card-generation-type"
        ]
      }
    },
...
    "/achievement-criteria": {
      "post": {
        "operationId": "AchievementCriteriaController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAchievementCriteriaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "achievement-criteria"
        ]
      },
      "get": {
        "operationId": "AchievementCriteriaController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "achievement-criteria"
        ]
      }
    },
    "/achievement-criteria/{id}": {
      "get": {
        "operationId": "AchievementCriteriaController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "achievement-criteria"
        ]
      },
      "patch": {
        "operationId": "AchievementCriteriaController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAchievementCriteriaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "achievement-criteria"
        ]
      },
      "delete": {
        "operationId": "AchievementCriteriaController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "achievement-criteria"
        ]
      }
    },
...
  },
  "info": {
    "title": "Nest API",
    "description": "API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email for registration",
            "example": "user@example.com"
          },
          "username": {
            "type": "string",
            "description": "Unique username",
            "example": "john_doe"
          },
          "password": {
            "type": "string",
            "description": "Password (will be hashed before saving)",
            "example": "MyStrongP@ssw0rd"
          },
          "fullName": {
            "type": "string",
            "description": "Full name of the user",
            "example": "John Doe"
          },
          "phone": {
            "type": "string",
            "description": "Optional phone number",
            "example": "+1-555-1234567"
          },
          "profilePhoto": {
            "type": "string",
            "description": "Optional profile photo URL",
            "example": "https://example.com/avatar.jpg"
          },
          "roles": {
            "description": "Initial roles for the user",
            "example": [
              "user"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "email",
          "username",
          "password",
          "fullName"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email for login",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "MyStrongP@ssw0rd"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "CreateEventTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the event type",
            "example": "Conference"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly slug for the event type (auto-generated from name if not provided)",
            "example": "conference"
          },
          "description": {
            "type": "string",
            "description": "Description of the event type",
            "example": "Business conferences and seminars"
          },
          "status": {
            "type": "string",
            "description": "Status of the event type",
            "example": "active",
            "enum": [
              "active",
              "inactive"
            ]
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateEventTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the event type",
            "example": "Conference"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly slug for the event type",
            "example": "conference"
          },
          "description": {
            "type": "string",
            "description": "Description of the event type",
            "example": "Business conferences and seminars"
          },
          "status": {
            "type": "string",
            "description": "Status of the event type",
            "example": "active",
            "enum": [
              "active",
              "inactive"
            ]
          }
        }
      },
...
      "CreateEligibilityDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the eligibility",
            "example": "Adult"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly slug for the eligibility (auto-generated from name if not provided)",
            "example": "adult"
          },
          "status": {
            "type": "string",
            "description": "Status of the eligibility",
            "example": "active",
            "enum": [
              "active",
              "inactive"
            ]
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateEligibilityDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the eligibility",
            "example": "Adult"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly slug for the eligibility",
            "example": "adult"
          },
          "status": {
            "type": "string",
            "description": "Status of the eligibility",
            "example": "active",
            "enum": [
              "active",
              "inactive"
            ]
          }
        }
      },
...
      "UpdateEventDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the event",
            "example": "Tech Conference 2024"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly slug for the event",
            "example": "tech-conference-2024"
          },
          "category_id": {
            "type": "string",
            "description": "Category ID (references Categories)",
            "example": "665f1a4b2c3d4e5f6a7b8d0e"
          },
          "event_type_id": {
            "type": "string",
            "description": "Event Type ID (references EventTypes)",
            "example": "665f1a4b2c3d4e5f6a7b8d0e"
          },
          "start_date": {
            "type": "string",
            "description": "Start date/time for the event (ISO string)",
            "example": "2026-03-01T09:00:00Z"
          },
          "end_date": {
            "type": "string",
            "description": "End date/time for the event (ISO string)",
            "example": "2026-03-01T17:00:00Z"
          },
          "eligibility_id": {
            "type": "string",
            "description": "Eligibility ID (references Eligibilities)",
            "example": "665f1a4b2c3d4e5f6a7b8d0e"
          },
          "winning_condition": {
            "type": "string",
            "description": "Winning condition (text input)",
            "example": "First to complete 5 rounds"
          },
          "game_type_id": {
            "type": "string",
            "description": "Game Type ID (references Game Types)",
            "example": "665f1a4b2c3d4e5f6a7b8d0e"
          },
          "max_participants": {
            "type": "number",
            "description": "Maximum number of participants",
            "example": 100
          },
          "reward_id": {
            "type": "string",
            "description": "Reward ID (references Rewards module)",
            "example": "665f1a4b2c3d4e5f6a7b8d0e"
          },
          "rewards_value": {
            "type": "number",
            "description": "Reward value (points, amount, etc.)",
            "example": 500
          },
          "distribution": {
            "type": "string",
            "description": "Rewards distribution method",
            "example": "top-3"
          },
          "status": {
            "type": "string",
            "description": "Status of the event",
            "example": "active",
            "enum": [
              "active",
              "inactive"
            ]
          },
          "description": {
            "type": "string",
            "description": "Event description",
            "example": "Annual tech conference"
          }
        }
      },
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the category",
            "example": "Lodging"
          },
          "description": {
            "type": "string",
            "description": "Short description of the category",
            "example": "Accommodation and lodging facilities"
          },
          "status": {
            "type": "string",
            "description": "Status of the category",
            "example": "active",
            "enum": [
              "active",
              "inactive"
            ]
          },
          "visibility_type": {
            "type": "string",
            "description": "Visibility type of the category",
            "example": "FREE",
            "enum": [
              "FREE",
              "PREMIUM",
              "ORG_ONLY"
            ]
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the category",
            "example": "Lodging"
          },
          "description": {
            "type": "string",
            "description": "Short description of the category",
            "example": "Accommodation and lodging facilities"
          },
          "status": {
            "type": "string",
            "description": "Status of the category",
            "example": "active",
            "enum": [
              "active",
              "inactive"
            ]
          },
          "visibility_type": {
            "type": "string",
            "description": "Visibility type of the category",
            "example": "FREE",
            "enum": [
              "FREE",
              "PREMIUM",
              "ORG_ONLY"
            ]
          }
        }
      },
...
      "CreateCardGenerationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name",
            "example": "Standard"
          },
          "slug": {
            "type": "string",
            "description": "Slug (auto-generated from name if not provided)",
            "example": "standard"
          },
          "status": {
            "type": "string",
            "description": "Status",
            "example": "active",
            "enum": [
              "active",
              "inactive"
            ]
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCardGenerationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name",
            "example": "Standard"
          },
          "slug": {
            "type": "string",
            "description": "Slug",
            "example": "standard"
          },
          "status": {
            "type": "string",
            "description": "Status",
            "example": "active",
            "enum": [
              "active",
              "inactive"
            ]
          }
        }
      },
...
      "CreateAchievementCriteriaDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name",
            "example": "First Win"
          },
          "slug": {
            "type": "string",
            "description": "Slug (auto-generated from name if not provided)",
            "example": "first-win"
          },
          "status": {
            "type": "string",
            "description": "Status",
            "example": "active",
            "enum": [
              "active",
              "inactive"
            ]
          }
        },
        "required": [
          "name"
        ]
      },
...
    }
  }
}